Split PipelinePage and PipelineResults
